requirements/production.txt
gunicorn==20.1.0.0
django==4.0
djangorestframework==3.13.1
drf-flex-fields==0.9.7
djangorestframework-simplejwt==5.0.0
djangorestframework-camel-case==1.3.0
django-cors-headers==3.11.0
django-jazzmin==2.4.9
psycopg2==2.9.3
requests==2.27.1
# development req
drf-spectacular==0.21.2
django_advance_dumpdata
django-user-agents==0.4.0
ipy==1.1
phonenumbers==8.12.44
beautifulsoup4==4.10.0
sentry-sdk
pillow==9.0.0
django-filter
djangorestframework-api-key==2.2.0
readtime
django-dbbackup==4.0.0b0
cryptography==3.4.7
binance-pay-connector==1.0.0
